Move argument collection to common library
The static analysis tool needs this functionality to be exposed (rather
than just inside the instrumenter). This commit moves the argument
collection code into the common section, and has the previous methods
act as a proxy to the new location to preserve backwards compatibility.
