+++ tesla/instrumenter/Instrumentation.cpp
@@ -520,51 +520,6 @@ Constant* tesla::TeslaContext(AutomatonDescription::Context Context,
 }
 
 
-Value* tesla::GetArgumentValue(Value* Param, const Argument& ArgDescrip,
-                               IRBuilder<>& Builder, bool AtAssertionSite) {
-
-  switch (ArgDescrip.type()) {
-  case Argument::Constant:
-  case Argument::Any:
-  case Argument::Variable:
-    return Param;
-
-  case Argument::Indirect:
-    // At the assertion site, we don't do the indirection thing: we are
-    // referring to local values.
-    if (AtAssertionSite)
-      return Param;
-
-    Param = Builder.CreateLoad(Param);
-    return GetArgumentValue(Param, ArgDescrip.indirection(), Builder);
-
-  case Argument::Field: {
-    // We only need to do the indirect lookup via struct fields at the
-    // assertion site; otherwise, we just care about local values.
-    if (!AtAssertionSite)
-      return Param;
-
-    const StructField& Field = ArgDescrip.field();
-    const Argument& BaseArg = Field.base();
-    auto *Base = GetArgumentValue(Param, BaseArg, Builder, AtAssertionSite);
-
-    string Name = (
-      (Base->hasName()
-        ? Base->getName()
-        : StringRef(BaseName(BaseArg))
-      )
-      + "." + Field.name()
-    ).str();
-
-    Param = Builder.CreateConstInBoundsGEP2_32(Base, 0, Field.index());
-    Param = Builder.CreateLoad(Param, Name);
-
-    return Param;
-  }
-  }
-}
-
-
 Value* tesla::ConstructKey(IRBuilder<>& Builder, Module& M,
                            ArrayRef<Value*> Args, uint32_t FreeMask) {
